% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{avail}
\alias{avail}
\title{Availability (1 - unavailability / total)}
\usage{
avail(unavailTime, totalTime)
}
\arguments{
\item{unavailTime}{Numeric scalar or numeric vector of unavailable times.}

\item{totalTime}{Numeric scalar or numeric vector of total times (same units as unavailTime).}
}
\value{
Numeric scalar: availability for the period (between 0 and 1).
}
\description{
Availability is the proportion of time a system is in a functioning condition.
This function computes availability given unavailable times and total times.
}
\examples{
avail(100, 1000)
avail(c(5,10), c(500,600))
}
