% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{lda}
\alias{lda}
\title{Launch the Life Data Analysis Tutorial}
\usage{
lda()
}
\value{
This function does not return a value.
}
\description{
This function launches an interactive tutorial for life data analysis.
}
\examples{
if (interactive()) {
  lda()
}
}
\seealso{
\url{https://paulgovan.github.io/ReliaLearnR/}
}
