% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram}
\alias{ram}
\title{Launch the RAM Analysis Tutorial}
\usage{
ram()
}
\value{
This function does not return a value.
}
\description{
This function launches an interactive tutorial on Reliability, Availability,
and Maintainability (RAM) analysis.
}
\examples{
if (interactive()) {
  ram()
}
}
\seealso{
\url{https://paulgovan.github.io/ReliaLearnR/}
}
