% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbd.R
\name{rbd}
\alias{rbd}
\title{Launch the Reliability Block Diagram Tutorial}
\usage{
rbd()
}
\value{
This function does not return a value.
}
\description{
This function launches an interactive tutorial on Reliability Block Diagrams
and system reliability, covering series, parallel, mixed, and k-out-of-n
configurations as well as an introduction to Fault Tree Analysis.
}
\examples{
if (interactive()) {
  rbd()
}
}
\seealso{
\url{https://paulgovan.github.io/ReliaLearnR/}
}
