% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliaLearnR-package.R
\docType{package}
\name{ReliaLearnR-package}
\alias{ReliaLearnR}
\alias{ReliaLearnR-package}
\title{ReliaLearnR: Learning Modules for Reliability Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Learning modules for reliability analysis including modules for Reliability, Availability, and Maintainability (RAM) Analysis, Life Data Analysis, and Reliability Testing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paulgovan.github.io/ReliaLearnR/}
  \item \url{https://github.com/paulgovan/ReliaLearnR}
  \item Report bugs at \url{https://github.com/paulgovan/ReliaLearnR/issues}
}

}
\author{
\strong{Maintainer}: Paul Govan \email{paul.govan2@gmail.com} (\href{https://orcid.org/0000-0002-1821-8492}{ORCID}) [copyright holder]

}
\keyword{internal}
