% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs.R
\name{rs}
\alias{rs}
\title{Launch the Repairable Systems Analysis Tutorial}
\usage{
rs()
}
\value{
This function does not return a value.
}
\description{
This function launches an interactive tutorial on Repairable Systems Analysis.
}
\examples{
if (interactive()) {
  rs()
}
}
\seealso{
\url{https://paulgovan.github.io/ReliaLearnR/}
}
