% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{rt}
\alias{rt}
\title{Launch the Reliability Testing Tutorial}
\usage{
rt()
}
\value{
This function does not return a value.
}
\description{
This function launches the Reliability Testing tutorial
}
\examples{
if (interactive()) {
  rt()
}
}
\seealso{
\url{https://paulgovan.github.io/ReliaLearnR/}
}
