% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeibullR.learnr.R
\name{WeibullR.learnr}
\alias{WeibullR.learnr}
\title{An Interactive Introduction to Life Data Analysis}
\usage{
WeibullR.learnr()
}
\description{
WeibullR.learnr is an interactive introduction to Life Data Analysis.
}
\examples{
if (interactive()) {
  WeibullR.learnr()
}
}
\seealso{
\url{http://paulgovan.github.io/WeibullR.learnr/}
}
